Секундный отсчёт и исправление типа данных в сумме сделки

+++ BinanceClient/BinanceClient/BinanceInfo.cs
@@ -11,10 +11,10 @@ namespace BinanceClient
         public long Id { get; set; }
         public DateTime Time { get; set; }
         public string Symbol { get; set; }
-        public int TradeQuantity { get; set; }
+        public decimal TradeQuantity { get; set; }
         public decimal RatePrice { get; set; }
 
-        public BinanceInfo(long id, DateTime time, string symbol, int tradeQuantity, decimal ratePrice)
+        public BinanceInfo(long id, DateTime time, string symbol, decimal tradeQuantity, decimal ratePrice)
         {
             Id = id;
             Time = time;
